requirements.txt
matplotlib==3.2.1
numpy==1.18.3
opencv-python==4.5.1.48
Pillow==8.1.0